Remove HeatgridLoader and make ParticleSystemVisualizer asynchronous
This is a bit of an unconventional update. Normally I would keep my code cleaner than this. There's still the issue of how you could make 2D grids, that the ParticleSystemVisualizer is bloated, that IVisualizer is NOT a very good contract as it basically allows anything... but I am done working on Heatgrids. It could be a lot better but I don't want to keep working on it that's already a little fraught with issues, since I could probably make something much better if I started from scratch. I will probably keep this package in experimental forever for that reason, however I did learn a lot, and it certainly paved the way for cubusky.com.ghosts. Lets me focus my efforts there.